Moved files and folders into a new src folder.
