+++ src/components/sections/about/Bio.tsx
