+++ src/components/sections/about/bio.ts
