+++ src/components/sections/greeting/TypeWriter.tsx
