+++ src/theme/layout/Layout.tsx
@@ -1,7 +1,7 @@
 import React, { FC, ReactNode } from "react";
 
 import type { AppProps } from "next/app";
-import Header from "../layout/Header";
+import Header from "./Header";
 import { Box } from "@chakra-ui/layout";
 import Footer from "./Footer";
